Minor refactor. Removing unused code.
