+++ FleetDaemon/App.config
@@ -23,7 +23,7 @@
       </basicHttpBinding>
     </bindings>
     <client>
-      <endpoint address="http://localhost:8733/Design_Time_Addresses/FleetServer/FleetService"
+      <endpoint address="http://HTCC-WIN1/FleetServer/FleetServer.FleetService.svc"
           binding="basicHttpBinding" bindingConfiguration="BasicHttpBinding_IFleetService"
           contract="IFleetService" name="BasicHttpBinding_IFleetService" />
     </client>
